requirements.txt
aiohttp
aiojobs
aiohttp_socks
colorama
discord.py